// frontend/src/api/models/handlerWarehouseResponseStatus.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type HandlerWarehouseResponseStatus =
  (typeof HandlerWarehouseResponseStatus)[keyof typeof HandlerWarehouseResponseStatus]

export const HandlerWarehouseResponseStatus = {
  enabled: 'enabled',
  disabled: 'disabled',
} as const