// frontend/src/api/models/handlerWarehouseResponseType.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type HandlerWarehouseResponseType =
  (typeof HandlerWarehouseResponseType)[keyof typeof HandlerWarehouseResponseType]

export const HandlerWarehouseResponseType = {
  normal: 'normal',
  virtual: 'virtual',
  transit: 'transit',
} as const